create timeExtractor util, move axios post and delete to actions
