+++ src/actions/index.js
@@ -1,6 +1,3 @@
-import axios from 'axios';
-import formatTime from '../utils/constants';
-
 export const FETCH_PRODUCTS_PENDING = 'FETCH_PRODUCTS_PENDING';
 export const FETCH_PRODUCTS_SUCCESS = 'FETCH_PRODUCTS_SUCCESS';
 export const FETCH_PRODUCTS_ERROR = 'FETCH_PRODUCTS_ERROR';
@@ -85,35 +82,9 @@ function fetchProgress() {
   };
 }
 
-function postMeasurement(id, seconds) {
-  axios.post(
-    'http://localhost:3001/api/v1/measurements',
-    {
-      time: formatTime(seconds),
-      date: Date(),
-      sport_id: id,
-    },
-    {
-      headers: {
-        'content-type': 'application/json',
-        'Access-Control-Allow-Origin': '*',
-        'Access-Control-Allow-Methods': 'GET,PUT,POST,DELETE,PATCH',
-      },
-    },
-  )
-    .then(res => {
-      if (res.error) {
-        throw (res.error);
-      }
-      window.location.reload(false);
-      return res;
-    }).catch(error => error);
-}
-
 export {
   fetchProducts,
   fetchProductsError,
   addUsername,
   fetchProgress,
-  postMeasurement,
 };
